+++ src/kitconcept/volto/transforms.py
@@ -2,12 +2,14 @@ from copy import deepcopy
 from plone.restapi.behaviors import IBlocks
 from plone.restapi.deserializer.blocks import path2uid
 from plone.restapi.interfaces import IBlockFieldDeserializationTransformer
+from plone.restapi.interfaces import IBlockFieldSerializationTransformer
 from plone.restapi.serializer.blocks import uid_to_url
 from Products.CMFPlone.interfaces import IPloneSiteRoot
 from six import string_types
 from zope.component import adapter
 from zope.interface import implementer
 from zope.publisher.interfaces.browser import IBrowserRequest
+from zope.component import subscribers
 
 
 class NestedResolveUIDDeserializerBase(object):
@@ -24,8 +26,25 @@ class NestedResolveUIDDeserializerBase(object):
         self.context = context
         self.request = request
 
+    def _transform(self, block):
+        """ this mutates the object directly """
+
+        block_type = block.get("@type", "")
+        handlers = []
+        for h in subscribers(
+            (self.context, self.request), IBlockFieldDeserializationTransformer
+        ):
+            if h.block_type == block_type or h.block_type is None:
+                h.blockid = block.get("id", None)
+                handlers.append(h)
+
+        for handler in sorted(handlers, key=lambda h: h.order):
+            block = handler(block)
+
+        return block
+
     def __call__(self, block):
-        for column_name in ["columns", "hrefList"]:
+        for column_name in ["columns", "hrefList", "slides"]:
             column_field = block.get(column_name, [])
             if block.get(column_name, False):
                 for index, item in enumerate(column_field):
@@ -56,6 +75,11 @@ class NestedResolveUIDDeserializerBase(object):
                                     path2uid(context=self.context, link=linkitem)
                                     for linkitem in link
                                 ]
+
+                    # Support for applying transforms to the subblocks in volto-blocks-grid
+                    # TODO: It's the upper code needed any longer?
+                    self._transform(item)
+
         return block
 
 
@@ -79,8 +103,25 @@ class NestedResolveUIDSerializerBase(object):
         self.context = context
         self.request = request
 
+    def _transform(self, block):
+        """ this mutates the object directly """
+
+        block_type = block.get("@type", "")
+        handlers = []
+        for h in subscribers(
+            (self.context, self.request), IBlockFieldSerializationTransformer
+        ):
+            if h.block_type == block_type or h.block_type is None:
+                h.blockid = block.get("id", None)
+                handlers.append(h)
+
+        for handler in sorted(handlers, key=lambda h: h.order):
+            block = handler(block)
+
+        return block
+
     def __call__(self, value):
-        for column_name in ["columns", "hrefList"]:
+        for column_name in ["columns", "hrefList", "slides"]:
             column_field = value.get(column_name, [])
             if value.get(column_name, False):
                 for index, item in enumerate(column_field):
@@ -108,6 +149,10 @@ class NestedResolveUIDSerializerBase(object):
                                     value[column_name][index][field] = [
                                         uid_to_url(linkitem) for linkitem in link
                                     ]
+
+                    # Support for applying transforms to the subblocks in volto-blocks-grid
+                    # TODO: It's the upper code needed any longer?
+                    self._transform(item)
 
         return value